feat: add update script

+++ install.sh
@@ -2,9 +2,16 @@
 #
 # Install script
 #
+WHITE="\e[0m"
+GREEN="\e[32;1m"
+
 destination="${1:-"${HOME}/.local/bin"}"
 
-# Setup pm before installing
+# Install templates and backends
 ./pm install-hook
 
-command cp -i pm "${destination}/pm"
+if [[ ! -f "${destination}/pm" ]]; then
+    command cp -i pm "${destination}/pm"
+fi
+
+echo -e "${GREEN}✓${WHITE} pm is ready to use"
